+++ src/pages/Todos/TodoDetail/TodoDetail.tsx
@@ -1,7 +1,9 @@
 import { useParams } from 'react-router-dom';
-import { useQuery, useMutation } from 'react-query';
-import { Todo, todoApi, updateTodoApi } from '../../../api/todoApi';
-import { useRef, useState } from 'react';
+import { useQuery } from 'react-query';
+import { Todo, todoApi } from '../../../api/todoApi';
+import { useState } from 'react';
+import UpdateTodo from './UpdateTodo';
+import ShowTodo from './ShowTodo';
 function TodoDetail() {
   const { id } = useParams();
   const authToken = localStorage.getItem('authtoken');
@@ -11,26 +13,8 @@ function TodoDetail() {
   const [isUpdate, setIsUpdate] = useState(false);
 
   const todo: Todo | undefined = data?.data?.data ?? undefined;
-  //todo 추가하기
+  //todo 추가하기 기능
   //   useMutation()
-
-  //TODO : 수정하기 영역은 컴포넌트화 하고 수정 useMutate 하는 코드도 같이 컴포넌트화
-  //   interface
-
-  //todo 수정하기
-  const titleRef = useRef<HTMLInputElement>(null);
-  const contentRef = useRef<HTMLInputElement>(null);
-
-  const todoUpdateMutation = useMutation(updateTodoApi, {
-    onSuccess: data => {
-      console.log(data);
-      alert('업데이트 완료!');
-    },
-    onError: e => {
-      console.error(e);
-      alert('업데이트 실패!');
-    },
-  });
 
   //TODO : error 시에 alert 및 뒤로가기
   return (
@@ -40,30 +24,9 @@ function TodoDetail() {
       ) : (
         <>
           {!isUpdate ? (
-            <div>
-              <p>{todo!.title}</p>
-              <p>{todo!.content}</p>
-              <button onClick={() => setIsUpdate(true)}>수정하기</button>
-            </div>
+            <ShowTodo todo={todo!} setIsUpdate={setIsUpdate}></ShowTodo>
           ) : (
-            <div>
-              <input type="text" defaultValue={todo!.title} ref={titleRef} />
-              <input type="text" defaultValue={todo!.content} ref={contentRef} />
-              수정모드
-              <button
-                onClick={() => {
-                  todoUpdateMutation.mutate({
-                    authToken: authToken!,
-                    id: todo!.id,
-                    title: titleRef.current!.value,
-                    content: contentRef.current!.value,
-                  });
-                }}
-              >
-                수정완료
-              </button>
-              <button onClick={() => setIsUpdate(false)}>취소하기</button>
-            </div>
+            <UpdateTodo todo={todo!} setIsUpdate={setIsUpdate}></UpdateTodo>
           )}
         </>
       )}
